add single report reducers and actions
