+++ frontend/src/redux/actions/ActionTypes.js
@@ -1,4 +1,4 @@
 export const UPDATE_REPORT_NAME = 'UPDATE_REPORT_NAME'
 export const UPDATE_REPORT_DESCRIPTION = 'UPDATE_REPORT_DESCRIPTION'
-export const UDPATE_REPORT_CARDS = 'UPDATE_REPORT_CARDS'
+export const UPDATE_REPORT_CARDS = 'UPDATE_REPORT_CARDS'
 export const SAVE_REPORT = 'SAVE_REPORT'
